feat: add path finder as domain object
