+++ src/main/java/nextstep/subway/domain/PathFinder.java
@@ -1,0 +1,10 @@
+package nextstep.subway.domain;
+
+import java.util.List;
+import nextstep.subway.applicaion.dto.PathRequest;
+import nextstep.subway.applicaion.dto.PathResponse;
+
+public interface PathFinder {
+    PathResponse findPath(PathRequest pathRequest, List<Line> lines);
+    void validateRequest(PathRequest request, List<Line> lines);
+}
